refactor(server): move tokens-data adapter into features/token-prices/data
Adapter was previously at shared/clients/tokens-data/ but has exactly
one consumer (token-prices). Per the architecture review, "shared"
implies multiple consumers; one consumer means the code belongs with
its feature. Generated types move along with the adapter.

DI register and both deprecation bridges (token-prices, tokens-meta)
updated to the new import paths. Turbo outputs for codegen updated to
match the new generated file location.
