+++ apps/server/src/features/token-prices/data/tokens-data-client.test.ts
@@ -1,6 +1,6 @@
 import "reflect-metadata";
 
-import { TokensDataClient, TokensDataError } from "./client";
+import { TokensDataClient, TokensDataError } from "./tokens-data-client";
 import { describe, expect, test } from "bun:test";
 
 class TestableClient extends TokensDataClient {
