+++ apps/server/src/features/token-prices/di/register.ts
@@ -1,6 +1,6 @@
 import { container } from "tsyringe";
 
-import { TokensDataPriceAdapter } from "../../../shared/clients/tokens-data/price-adapter";
+import { TokensDataPriceAdapter } from "../data/tokens-data-price-adapter";
 import { TOKEN_PRICE_SERVICE } from "./tokens";
 
 export function register() {
