+++ turbo.json
@@ -17,11 +17,11 @@
       "outputs": [
         "apps/server/openapi/**",
         "apps/server/src/features/uniswap-v3/data/gql/**",
-        "apps/server/src/shared/clients/tokens-data/generated.ts",
+        "apps/server/src/features/token-prices/data/tokens-data.generated.ts",
         "apps/tokens-data/generated/**",
         "apps/mobile/src/core/api-client/generated/**",
         "apps/mobile/src/features/uniswap-v3/data/generated/**",
-        "apps/mobile/src/core/tokens-data/**"
+        "apps/mobile/src/core/tokens-data/generated.d.ts"
       ],
       "env": ["GRAPH_API_KEY"]
     },
